pyproject.toml
[project]
name = "easy-ads"
version = "0.1.0"
description = "Creative Automation Pipeline for generating social media campaign assets using GenAI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "Pillow>=10.3.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "numpy>=1.26.0",
    "huggingface_hub>=0.20.0",
    "python-dotenv>=1.0.0",
    "replicate>=0.25.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["pipeline", "backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
